requirements.txt
tensorflow
sentence-transformers
yfinance
newspaper3k
gradio
pandas